refactor: split agent output handler modules
